refactor raptor
