+++ src/extensions/surveys.tsx
@@ -1,36 +1,18 @@
 import { PostHog } from '../posthog-core'
-import {
-    Survey,
-    SurveyAppearance,
-    SurveyQuestion,
-    SurveyQuestionBranchingType,
-    SurveyQuestionType,
-    SurveyType,
-} from '../posthog-surveys-types'
+import { Survey, SurveyType } from '../posthog-surveys-types'
 
 import { window as _window, document as _document } from '../utils/globals'
 import {
     style,
     defaultSurveyAppearance,
-    sendSurveyEvent,
-    dismissedSurveyEvent,
     createShadow,
     getContrastingTextColor,
-    SurveyContext,
-    getDisplayOrderQuestions,
     getSurveySeenKey,
 } from './surveys/surveys-utils'
 import * as Preact from 'preact'
 import { createWidgetShadow, createWidgetStyle } from './surveys-widget'
-import { useState, useEffect, useRef, useContext, useMemo } from 'preact/hooks'
-import { isNumber } from '../utils/type-utils'
-import { ConfirmationMessage } from './surveys/components/ConfirmationMessage'
-import {
-    OpenTextQuestion,
-    LinkQuestion,
-    RatingQuestion,
-    MultipleChoiceQuestion,
-} from './surveys/components/QuestionTypes'
+import { FeedbackWidget } from './surveys/components/FeedbackWidget'
+import { SurveyPopup } from './surveys/components/SurveyPopup'
 
 // We cast the types here which is dangerous but protected by the top level generateSurveys call
 const window = _window as Window & typeof globalThis
@@ -271,360 +253,360 @@ export function generateSurveys(posthog: PostHog) {
     }, 1000)
 }
 
-export function usePopupVisibility(
-    survey: Survey,
-    posthog: PostHog | undefined,
-    millisecondDelay: number,
-    isPreviewMode: boolean,
-    removeSurveyFromFocus: (id: string) => void
-) {
-    const [isPopupVisible, setIsPopupVisible] = useState(isPreviewMode || millisecondDelay === 0)
-    const [isSurveySent, setIsSurveySent] = useState(false)
+// export function usePopupVisibility(
+//     survey: Survey,
+//     posthog: PostHog | undefined,
+//     millisecondDelay: number,
+//     isPreviewMode: boolean,
+//     removeSurveyFromFocus: (id: string) => void
+// ) {
+//     const [isPopupVisible, setIsPopupVisible] = useState(isPreviewMode || millisecondDelay === 0)
+//     const [isSurveySent, setIsSurveySent] = useState(false)
 
-    useEffect(() => {
-        if (isPreviewMode || !posthog) {
-            return
-        }
+//     useEffect(() => {
+//         if (isPreviewMode || !posthog) {
+//             return
+//         }
 
-        const handleSurveyClosed = () => {
-            removeSurveyFromFocus(survey.id)
-            setIsPopupVisible(false)
-        }
+//         const handleSurveyClosed = () => {
+//             removeSurveyFromFocus(survey.id)
+//             setIsPopupVisible(false)
+//         }
 
-        const handleSurveySent = () => {
-            if (!survey.appearance?.displayThankYouMessage) {
-                removeSurveyFromFocus(survey.id)
-                setIsPopupVisible(false)
-            } else {
-                setIsSurveySent(true)
-                if (survey.appearance?.autoDisappear) {
-                    setTimeout(() => {
-                        removeSurveyFromFocus(survey.id)
-                        setIsPopupVisible(false)
-                    }, 5000)
-                }
-            }
-        }
+//         const handleSurveySent = () => {
+//             if (!survey.appearance?.displayThankYouMessage) {
+//                 removeSurveyFromFocus(survey.id)
+//                 setIsPopupVisible(false)
+//             } else {
+//                 setIsSurveySent(true)
+//                 if (survey.appearance?.autoDisappear) {
+//                     setTimeout(() => {
+//                         removeSurveyFromFocus(survey.id)
+//                         setIsPopupVisible(false)
+//                     }, 5000)
+//                 }
+//             }
+//         }
 
-        window.addEventListener('PHSurveyClosed', handleSurveyClosed)
-        window.addEventListener('PHSurveySent', handleSurveySent)
+//         window.addEventListener('PHSurveyClosed', handleSurveyClosed)
+//         window.addEventListener('PHSurveySent', handleSurveySent)
 
-        if (millisecondDelay > 0) {
-            const timeoutId = setTimeout(() => {
-                setIsPopupVisible(true)
-                window.dispatchEvent(new Event('PHSurveyShown'))
-                posthog.capture('survey shown', {
-                    $survey_name: survey.name,
-                    $survey_id: survey.id,
-                    $survey_iteration: survey.current_iteration,
-                    $survey_iteration_start_date: survey.current_iteration_start_date,
-                    sessionRecordingUrl: posthog.get_session_replay_url?.(),
-                })
-                localStorage.setItem(`lastSeenSurveyDate`, new Date().toISOString())
-            }, millisecondDelay)
+//         if (millisecondDelay > 0) {
+//             const timeoutId = setTimeout(() => {
+//                 setIsPopupVisible(true)
+//                 window.dispatchEvent(new Event('PHSurveyShown'))
+//                 posthog.capture('survey shown', {
+//                     $survey_name: survey.name,
+//                     $survey_id: survey.id,
+//                     $survey_iteration: survey.current_iteration,
+//                     $survey_iteration_start_date: survey.current_iteration_start_date,
+//                     sessionRecordingUrl: posthog.get_session_replay_url?.(),
+//                 })
+//                 localStorage.setItem(`lastSeenSurveyDate`, new Date().toISOString())
+//             }, millisecondDelay)
 
-            return () => {
-                clearTimeout(timeoutId)
-                window.removeEventListener('PHSurveyClosed', handleSurveyClosed)
-                window.removeEventListener('PHSurveySent', handleSurveySent)
-            }
-        } else {
-            setIsPopupVisible(true)
-            window.dispatchEvent(new Event('PHSurveyShown'))
-            posthog.capture('survey shown', {
-                $survey_name: survey.name,
-                $survey_id: survey.id,
-                $survey_iteration: survey.current_iteration,
-                $survey_iteration_start_date: survey.current_iteration_start_date,
-                sessionRecordingUrl: posthog.get_session_replay_url?.(),
-            })
-            localStorage.setItem(`lastSeenSurveyDate`, new Date().toISOString())
+//             return () => {
+//                 clearTimeout(timeoutId)
+//                 window.removeEventListener('PHSurveyClosed', handleSurveyClosed)
+//                 window.removeEventListener('PHSurveySent', handleSurveySent)
+//             }
+//         } else {
+//             setIsPopupVisible(true)
+//             window.dispatchEvent(new Event('PHSurveyShown'))
+//             posthog.capture('survey shown', {
+//                 $survey_name: survey.name,
+//                 $survey_id: survey.id,
+//                 $survey_iteration: survey.current_iteration,
+//                 $survey_iteration_start_date: survey.current_iteration_start_date,
+//                 sessionRecordingUrl: posthog.get_session_replay_url?.(),
+//             })
+//             localStorage.setItem(`lastSeenSurveyDate`, new Date().toISOString())
 
-            return () => {
-                window.removeEventListener('PHSurveyClosed', handleSurveyClosed)
-                window.removeEventListener('PHSurveySent', handleSurveySent)
-            }
-        }
-    }, [])
+//             return () => {
+//                 window.removeEventListener('PHSurveyClosed', handleSurveyClosed)
+//                 window.removeEventListener('PHSurveySent', handleSurveySent)
+//             }
+//         }
+//     }, [])
 
-    return { isPopupVisible, isSurveySent, setIsPopupVisible }
-}
+//     return { isPopupVisible, isSurveySent, setIsPopupVisible }
+// }
 
-export function SurveyPopup({
-    survey,
-    forceDisableHtml,
-    posthog,
-    style,
-    previewPageIndex,
-    removeSurveyFromFocus,
-}: {
-    survey: Survey
-    forceDisableHtml?: boolean
-    posthog?: PostHog
-    style?: React.CSSProperties
-    previewPageIndex?: number | undefined
-    removeSurveyFromFocus: (id: string) => void
-}) {
-    const isPreviewMode = Number.isInteger(previewPageIndex)
-    // NB: The client-side code passes the millisecondDelay in seconds, but setTimeout expects milliseconds, so we multiply by 1000
-    const surveyPopupDelayMilliseconds = survey.appearance?.surveyPopupDelaySeconds
-        ? survey.appearance.surveyPopupDelaySeconds * 1000
-        : 0
-    const { isPopupVisible, isSurveySent, setIsPopupVisible } = usePopupVisibility(
-        survey,
-        posthog,
-        surveyPopupDelayMilliseconds,
-        isPreviewMode,
-        removeSurveyFromFocus
-    )
-    const shouldShowConfirmation = isSurveySent || previewPageIndex === survey.questions.length
-    const confirmationBoxLeftStyle = style?.left && isNumber(style?.left) ? { left: style.left - 40 } : {}
+// export function SurveyPopup({
+//     survey,
+//     forceDisableHtml,
+//     posthog,
+//     style,
+//     previewPageIndex,
+//     removeSurveyFromFocus,
+// }: {
+//     survey: Survey
+//     forceDisableHtml?: boolean
+//     posthog?: PostHog
+//     style?: React.CSSProperties
+//     previewPageIndex?: number | undefined
+//     removeSurveyFromFocus: (id: string) => void
+// }) {
+//     const isPreviewMode = Number.isInteger(previewPageIndex)
+//     // NB: The client-side code passes the millisecondDelay in seconds, but setTimeout expects milliseconds, so we multiply by 1000
+//     const surveyPopupDelayMilliseconds = survey.appearance?.surveyPopupDelaySeconds
+//         ? survey.appearance.surveyPopupDelaySeconds * 1000
+//         : 0
+//     const { isPopupVisible, isSurveySent, setIsPopupVisible } = usePopupVisibility(
+//         survey,
+//         posthog,
+//         surveyPopupDelayMilliseconds,
+//         isPreviewMode,
+//         removeSurveyFromFocus
+//     )
+//     const shouldShowConfirmation = isSurveySent || previewPageIndex === survey.questions.length
+//     const confirmationBoxLeftStyle = style?.left && isNumber(style?.left) ? { left: style.left - 40 } : {}
 
-    if (isPreviewMode) {
-        style = style || {}
-        style.left = 'unset'
-        style.right = 'unset'
-        style.transform = 'unset'
-    }
+//     if (isPreviewMode) {
+//         style = style || {}
+//         style.left = 'unset'
+//         style.right = 'unset'
+//         style.transform = 'unset'
+//     }
 
-    return isPopupVisible ? (
-        <SurveyContext.Provider
-            value={{
-                isPreviewMode,
-                previewPageIndex: previewPageIndex,
-                handleCloseSurveyPopup: () => {
-                    removeSurveyFromFocus(survey.id)
-                    dismissedSurveyEvent(survey, posthog, isPreviewMode)
-                },
-            }}
-        >
-            {!shouldShowConfirmation ? (
-                <Questions
-                    survey={survey}
-                    forceDisableHtml={!!forceDisableHtml}
-                    posthog={posthog}
-                    styleOverrides={style}
-                    removeSurveyFromFocus={removeSurveyFromFocus}
-                />
-            ) : (
-                <ConfirmationMessage
-                    header={survey.appearance?.thankYouMessageHeader || 'Thank you!'}
-                    description={survey.appearance?.thankYouMessageDescription || ''}
-                    forceDisableHtml={!!forceDisableHtml}
-                    contentType={survey.appearance?.thankYouMessageDescriptionContentType}
-                    appearance={survey.appearance || defaultSurveyAppearance}
-                    styleOverrides={{ ...style, ...confirmationBoxLeftStyle }}
-                    onClose={() => setIsPopupVisible(false)}
-                />
-            )}
-        </SurveyContext.Provider>
-    ) : (
-        <></>
-    )
-}
+//     return isPopupVisible ? (
+//         <SurveyContext.Provider
+//             value={{
+//                 isPreviewMode,
+//                 previewPageIndex: previewPageIndex,
+//                 handleCloseSurveyPopup: () => {
+//                     removeSurveyFromFocus(survey.id)
+//                     dismissedSurveyEvent(survey, posthog, isPreviewMode)
+//                 },
+//             }}
+//         >
+//             {!shouldShowConfirmation ? (
+//                 <Questions
+//                     survey={survey}
+//                     forceDisableHtml={!!forceDisableHtml}
+//                     posthog={posthog}
+//                     styleOverrides={style}
+//                     removeSurveyFromFocus={removeSurveyFromFocus}
+//                 />
+//             ) : (
+//                 <ConfirmationMessage
+//                     header={survey.appearance?.thankYouMessageHeader || 'Thank you!'}
+//                     description={survey.appearance?.thankYouMessageDescription || ''}
+//                     forceDisableHtml={!!forceDisableHtml}
+//                     contentType={survey.appearance?.thankYouMessageDescriptionContentType}
+//                     appearance={survey.appearance || defaultSurveyAppearance}
+//                     styleOverrides={{ ...style, ...confirmationBoxLeftStyle }}
+//                     onClose={() => setIsPopupVisible(false)}
+//                 />
+//             )}
+//         </SurveyContext.Provider>
+//     ) : (
+//         <></>
+//     )
+// }
 
-export function Questions({
-    survey,
-    forceDisableHtml,
-    posthog,
-    styleOverrides,
-    removeSurveyFromFocus,
-}: {
-    survey: Survey
-    forceDisableHtml: boolean
-    posthog?: PostHog
-    styleOverrides?: React.CSSProperties
-    removeSurveyFromFocus: (id: string) => void
-}) {
-    const textColor = getContrastingTextColor(
-        survey.appearance?.backgroundColor || defaultSurveyAppearance.backgroundColor
-    )
-    const [questionsResponses, setQuestionsResponses] = useState({})
-    const { isPreviewMode, previewPageIndex } = useContext(SurveyContext)
-    const [currentQuestionIndex, setCurrentQuestionIndex] = useState(previewPageIndex || 0)
-    const surveyQuestions = useMemo(() => getDisplayOrderQuestions(survey), [survey])
+// export function Questions({
+//     survey,
+//     forceDisableHtml,
+//     posthog,
+//     styleOverrides,
+//     removeSurveyFromFocus,
+// }: {
+//     survey: Survey
+//     forceDisableHtml: boolean
+//     posthog?: PostHog
+//     styleOverrides?: React.CSSProperties
+//     removeSurveyFromFocus: (id: string) => void
+// }) {
+//     const textColor = getContrastingTextColor(
+//         survey.appearance?.backgroundColor || defaultSurveyAppearance.backgroundColor
+//     )
+//     const [questionsResponses, setQuestionsResponses] = useState({})
+//     const { isPreviewMode, previewPageIndex } = useContext(SurveyContext)
+//     const [currentQuestionIndex, setCurrentQuestionIndex] = useState(previewPageIndex || 0)
+//     const surveyQuestions = useMemo(() => getDisplayOrderQuestions(survey), [survey])
 
-    // Sync preview state
-    useEffect(() => {
-        setCurrentQuestionIndex(previewPageIndex ?? 0)
-    }, [previewPageIndex])
+//     // Sync preview state
+//     useEffect(() => {
+//         setCurrentQuestionIndex(previewPageIndex ?? 0)
+//     }, [previewPageIndex])
 
-    const onNextButtonClick = ({
-        res,
-        originalQuestionIndex,
-        displayQuestionIndex,
-    }: {
-        res: string | string[] | number | null
-        originalQuestionIndex: number
-        displayQuestionIndex: number
-    }) => {
-        if (!posthog) {
-            return
-        }
+//     const onNextButtonClick = ({
+//         res,
+//         originalQuestionIndex,
+//         displayQuestionIndex,
+//     }: {
+//         res: string | string[] | number | null
+//         originalQuestionIndex: number
+//         displayQuestionIndex: number
+//     }) => {
+//         if (!posthog) {
+//             return
+//         }
 
-        const responseKey =
-            originalQuestionIndex === 0 ? `$survey_response` : `$survey_response_${originalQuestionIndex}`
+//         const responseKey =
+//             originalQuestionIndex === 0 ? `$survey_response` : `$survey_response_${originalQuestionIndex}`
 
-        setQuestionsResponses({ ...questionsResponses, [responseKey]: res })
+//         setQuestionsResponses({ ...questionsResponses, [responseKey]: res })
 
-        const nextStep = posthog.getNextSurveyStep(survey, displayQuestionIndex, res)
-        if (nextStep === SurveyQuestionBranchingType.ConfirmationMessage) {
-            removeSurveyFromFocus(survey.id)
-            sendSurveyEvent({ ...questionsResponses, [responseKey]: res }, survey, posthog)
-        } else {
-            setCurrentQuestionIndex(nextStep)
-        }
-    }
+//         const nextStep = posthog.getNextSurveyStep(survey, displayQuestionIndex, res)
+//         if (nextStep === SurveyQuestionBranchingType.ConfirmationMessage) {
+//             removeSurveyFromFocus(survey.id)
+//             sendSurveyEvent({ ...questionsResponses, [responseKey]: res }, survey, posthog)
+//         } else {
+//             setCurrentQuestionIndex(nextStep)
+//         }
+//     }
 
-    return (
-        <form
-            className="survey-form"
-            style={{
-                color: textColor,
-                borderColor: survey.appearance?.borderColor,
-                ...styleOverrides,
-            }}
-        >
-            {surveyQuestions.map((question, displayQuestionIndex) => {
-                const { originalQuestionIndex } = question
+//     return (
+//         <form
+//             className="survey-form"
+//             style={{
+//                 color: textColor,
+//                 borderColor: survey.appearance?.borderColor,
+//                 ...styleOverrides,
+//             }}
+//         >
+//             {surveyQuestions.map((question, displayQuestionIndex) => {
+//                 const { originalQuestionIndex } = question
 
-                const isVisible = isPreviewMode
-                    ? currentQuestionIndex === originalQuestionIndex
-                    : currentQuestionIndex === displayQuestionIndex
-                return (
-                    isVisible && (
-                        <div>
-                            {getQuestionComponent({
-                                question,
-                                forceDisableHtml,
-                                displayQuestionIndex,
-                                appearance: survey.appearance || defaultSurveyAppearance,
-                                onSubmit: (res) =>
-                                    onNextButtonClick({
-                                        res,
-                                        originalQuestionIndex,
-                                        displayQuestionIndex,
-                                    }),
-                            })}
-                        </div>
-                    )
-                )
-            })}
-        </form>
-    )
-}
+//                 const isVisible = isPreviewMode
+//                     ? currentQuestionIndex === originalQuestionIndex
+//                     : currentQuestionIndex === displayQuestionIndex
+//                 return (
+//                     isVisible && (
+//                         <div>
+//                             {getQuestionComponent({
+//                                 question,
+//                                 forceDisableHtml,
+//                                 displayQuestionIndex,
+//                                 appearance: survey.appearance || defaultSurveyAppearance,
+//                                 onSubmit: (res) =>
+//                                     onNextButtonClick({
+//                                         res,
+//                                         originalQuestionIndex,
+//                                         displayQuestionIndex,
+//                                     }),
+//                             })}
+//                         </div>
+//                     )
+//                 )
+//             })}
+//         </form>
+//     )
+// }
 
-export function FeedbackWidget({
-    survey,
-    forceDisableHtml,
-    posthog,
-    readOnly,
-    removeSurveyFromFocus,
-}: {
-    survey: Survey
-    forceDisableHtml?: boolean
-    posthog?: PostHog
-    readOnly?: boolean
-    removeSurveyFromFocus: (id: string) => void
-}): JSX.Element {
-    const [showSurvey, setShowSurvey] = useState(false)
-    const [styleOverrides, setStyle] = useState({})
-    const widgetRef = useRef<HTMLDivElement>(null)
+// export function FeedbackWidget({
+//     survey,
+//     forceDisableHtml,
+//     posthog,
+//     readOnly,
+//     removeSurveyFromFocus,
+// }: {
+//     survey: Survey
+//     forceDisableHtml?: boolean
+//     posthog?: PostHog
+//     readOnly?: boolean
+//     removeSurveyFromFocus: (id: string) => void
+// }): JSX.Element {
+//     const [showSurvey, setShowSurvey] = useState(false)
+//     const [styleOverrides, setStyle] = useState({})
+//     const widgetRef = useRef<HTMLDivElement>(null)
 
-    useEffect(() => {
-        if (readOnly || !posthog) {
-            return
-        }
+//     useEffect(() => {
+//         if (readOnly || !posthog) {
+//             return
+//         }
 
-        if (survey.appearance?.widgetType === 'tab') {
-            if (widgetRef.current) {
-                const widgetPos = widgetRef.current.getBoundingClientRect()
-                const style = {
-                    top: '50%',
-                    left: parseInt(`${widgetPos.right - 360}`),
-                    bottom: 'auto',
-                    borderRadius: 10,
-                    borderBottom: `1.5px solid ${survey.appearance?.borderColor || '#c9c6c6'}`,
-                }
-                setStyle(style)
-            }
-        }
-        if (survey.appearance?.widgetType === 'selector') {
-            const widget = document.querySelector(survey.appearance.widgetSelector || '')
-            widget?.addEventListener('click', () => {
-                setShowSurvey(!showSurvey)
-            })
-            widget?.setAttribute('PHWidgetSurveyClickListener', 'true')
-        }
-    }, [])
+//         if (survey.appearance?.widgetType === 'tab') {
+//             if (widgetRef.current) {
+//                 const widgetPos = widgetRef.current.getBoundingClientRect()
+//                 const style = {
+//                     top: '50%',
+//                     left: parseInt(`${widgetPos.right - 360}`),
+//                     bottom: 'auto',
+//                     borderRadius: 10,
+//                     borderBottom: `1.5px solid ${survey.appearance?.borderColor || '#c9c6c6'}`,
+//                 }
+//                 setStyle(style)
+//             }
+//         }
+//         if (survey.appearance?.widgetType === 'selector') {
+//             const widget = document.querySelector(survey.appearance.widgetSelector || '')
+//             widget?.addEventListener('click', () => {
+//                 setShowSurvey(!showSurvey)
+//             })
+//             widget?.setAttribute('PHWidgetSurveyClickListener', 'true')
+//         }
+//     }, [])
 
-    return (
-        <>
-            {survey.appearance?.widgetType === 'tab' && (
-                <div
-                    className="ph-survey-widget-tab"
-                    ref={widgetRef}
-                    onClick={() => !readOnly && setShowSurvey(!showSurvey)}
-                    style={{ color: getContrastingTextColor(survey.appearance.widgetColor) }}
-                >
-                    <div className="ph-survey-widget-tab-icon"></div>
-                    {survey.appearance?.widgetLabel || ''}
-                </div>
-            )}
-            {showSurvey && (
-                <SurveyPopup
-                    key={'feedback-widget-survey'}
-                    posthog={posthog}
-                    survey={survey}
-                    forceDisableHtml={forceDisableHtml}
-                    style={styleOverrides}
-                    removeSurveyFromFocus={removeSurveyFromFocus}
-                />
-            )}
-        </>
-    )
-}
+//     return (
+//         <>
+//             {survey.appearance?.widgetType === 'tab' && (
+//                 <div
+//                     className="ph-survey-widget-tab"
+//                     ref={widgetRef}
+//                     onClick={() => !readOnly && setShowSurvey(!showSurvey)}
+//                     style={{ color: getContrastingTextColor(survey.appearance.widgetColor) }}
+//                 >
+//                     <div className="ph-survey-widget-tab-icon"></div>
+//                     {survey.appearance?.widgetLabel || ''}
+//                 </div>
+//             )}
+//             {showSurvey && (
+//                 <SurveyPopup
+//                     key={'feedback-widget-survey'}
+//                     posthog={posthog}
+//                     survey={survey}
+//                     forceDisableHtml={forceDisableHtml}
+//                     style={styleOverrides}
+//                     removeSurveyFromFocus={removeSurveyFromFocus}
+//                 />
+//             )}
+//         </>
+//     )
+// }
 
-interface GetQuestionComponentProps {
-    question: SurveyQuestion
-    forceDisableHtml: boolean
-    displayQuestionIndex: number
-    appearance: SurveyAppearance
-    onSubmit: (res: string | string[] | number | null) => void
-}
+// interface GetQuestionComponentProps {
+//     question: SurveyQuestion
+//     forceDisableHtml: boolean
+//     displayQuestionIndex: number
+//     appearance: SurveyAppearance
+//     onSubmit: (res: string | string[] | number | null) => void
+// }
 
-const getQuestionComponent = ({
-    question,
-    forceDisableHtml,
-    displayQuestionIndex,
-    appearance,
-    onSubmit,
-}: GetQuestionComponentProps): JSX.Element => {
-    const questionComponents = {
-        [SurveyQuestionType.Open]: OpenTextQuestion,
-        [SurveyQuestionType.Link]: LinkQuestion,
-        [SurveyQuestionType.Rating]: RatingQuestion,
-        [SurveyQuestionType.SingleChoice]: MultipleChoiceQuestion,
-        [SurveyQuestionType.MultipleChoice]: MultipleChoiceQuestion,
-    }
+// const getQuestionComponent = ({
+//     question,
+//     forceDisableHtml,
+//     displayQuestionIndex,
+//     appearance,
+//     onSubmit,
+// }: GetQuestionComponentProps): JSX.Element => {
+//     const questionComponents = {
+//         [SurveyQuestionType.Open]: OpenTextQuestion,
+//         [SurveyQuestionType.Link]: LinkQuestion,
+//         [SurveyQuestionType.Rating]: RatingQuestion,
+//         [SurveyQuestionType.SingleChoice]: MultipleChoiceQuestion,
+//         [SurveyQuestionType.MultipleChoice]: MultipleChoiceQuestion,
+//     }
 
-    const commonProps = {
-        question,
-        forceDisableHtml,
-        appearance,
-        onSubmit,
-    }
+//     const commonProps = {
+//         question,
+//         forceDisableHtml,
+//         appearance,
+//         onSubmit,
+//     }
 
-    const additionalProps: Record<SurveyQuestionType, any> = {
-        [SurveyQuestionType.Open]: {},
-        [SurveyQuestionType.Link]: {},
-        [SurveyQuestionType.Rating]: { displayQuestionIndex },
-        [SurveyQuestionType.SingleChoice]: { displayQuestionIndex },
-        [SurveyQuestionType.MultipleChoice]: { displayQuestionIndex },
-    }
+//     const additionalProps: Record<SurveyQuestionType, any> = {
+//         [SurveyQuestionType.Open]: {},
+//         [SurveyQuestionType.Link]: {},
+//         [SurveyQuestionType.Rating]: { displayQuestionIndex },
+//         [SurveyQuestionType.SingleChoice]: { displayQuestionIndex },
+//         [SurveyQuestionType.MultipleChoice]: { displayQuestionIndex },
+//     }
 
-    const Component = questionComponents[question.type]
-    const componentProps = { ...commonProps, ...additionalProps[question.type] }
+//     const Component = questionComponents[question.type]
+//     const componentProps = { ...commonProps, ...additionalProps[question.type] }
 
-    return <Component {...componentProps} />
-}
+//     return <Component {...componentProps} />
+// }
